CV_acceleration/src/usb_2_0/requirements.txt
pyusb==1.3.1
libusb1==3.3.1
libusb-package==1.0.26.3
pywinusb==0.4.2
numpy==2.2.6
opencv-python==4.12.0.88